+++ evo_model/src/lib.rs
@@ -8,6 +8,6 @@ pub enum UserAction {
     DebugPrint,
     Exit,
     PlayToggle,
-    SelectCell { x: f32, y: f32 },
+    SelectCell { x: f64, y: f64 },
     SingleTick,
 }
